// frontend/src/api/generated/projects.ts
export interface RestProjectPriorityResponse {
  label: string;
  label_jp: string;
  value: string;
  weight: number;
}

export type RestProjectResponseStatus = typeof RestProjectResponseStatus[keyof typeof RestProjectResponseStatus];


export const RestProjectResponseStatus = {
  open: 'open',
  pending: 'pending',
  waiting_on_others: 'waiting_on_others',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface RestProjectTypeResponse {
  label: string;
  label_jp: string;
  value: string;
}

export interface RestProjectResponse {
  can_delete: boolean;
  can_update: boolean;
  created_at: number;
  /** @nullable */
  deleted_at: number | null;
  description: string;
  /** @nullable */
  end_date: string | null;
  goal: string;
  id: string;
  priority: RestProjectPriorityResponse;
  progress: number;
  /** @nullable */
  remaining_days: number | null;
  revision: number;
  /** @nullable */
  start_date: string | null;
  status: RestProjectResponseStatus;
  title: string;
  type: RestProjectTypeResponse;
  updated_at: number;
  user_id: string;
}
// ...
export interface RestProjectTaskStatusResponse {
  label?: string;
  label_jp?: string;
  value?: string;
}
// ...
export type GetProjectsStatus = typeof GetProjectsStatus[keyof typeof GetProjectsStatus];


export const GetProjectsStatus = {
  in_progress: 'in_progress',
  pending: 'pending',
  done: 'done',
  open: 'open',
  waiting_on_others: 'waiting_on_others',
} as const;
// ...
export type GetProjectsSortOrder = typeof GetProjectsSortOrder[keyof typeof GetProjectsSortOrder];


export const GetProjectsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;